+++ node/src/types/kernel/api/notebook/lsNotebooks/payload.d.ts
@@ -16,9 +16,14 @@
  */
 
 //#region content
-/* directories */
+/**
+ * List notebooks
+ */
+export interface IPayload {
+    /**
+     * Whether to return information about the flash card
+     */
+    flashcard?: boolean;
+}
 
-/* files */
-export * from "./payload";
-export * from "./response";
 //#endregion content